feature(plugin-nvidia): nvml refactoring and lib.rs modification for nvml
