+++ plugin-nvidia/Cargo.toml
@@ -19,5 +19,8 @@ nvml-wrapper-sys = { version = "0.8.0", optional = true }
 regex = { version = "1.10.6", optional = true }
 serde = { version = "1.0.210", features = ["derive"] }
 
+[dev-dependencies]
+tempfile = "3.10"
+
 [lints]
 workspace = true
