+++ plugin-nvidia/src/nvml/mod.rs
@@ -1,0 +1,4 @@
+pub mod device;
+pub mod features;
+pub mod metrics;
+pub mod probe;
